requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
blinker==1.9.0
blis==1.3.3
cachetools==7.0.6
catalogue==2.0.10
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cymem==2.0.13
decorator==5.2.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.136.1
ffmpeg-python==0.2.0
filelock==3.29.0
fsspec==2026.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
narwhals==2.20.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
openai-whisper==20250625
packaging==26.2
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
preshed==3.0.13
protobuf==7.34.1
pyarrow==24.0.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.26
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.6.2
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
whisper-openai==1.0.0
wrapt==2.1.2
